impl/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
aiofiles==23.2.1
    # via semantic-kernel
aiohappyeyeballs==2.3.4
    # via aiohttp
aiohttp==3.10.0
    # via semantic-kernel
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
asgiref==3.8.1
    # via openapi-core
attrs==24.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
azure-core==1.30.2
    # via
    #   azure-cosmos
    #   azure-identity
    #   azure-storage-blob
azure-cosmos==4.7.0
    # via -r .\requirements.in
azure-identity==1.17.1
    # via -r .\requirements.in
azure-storage-blob==12.21.0
    # via -r .\requirements.in
beautifulsoup4==4.12.3
    # via -r .\requirements.in
black==24.8.0
    # via -r .\requirements.in
certifi==2024.7.4
    # via
    #   -r .\requirements.in
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via prance
charset-normalizer==3.3.2
    # via
    #   httpie
    #   requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   httpie
    #   pytest
    #   tqdm
coverage[toml]==7.6.0
    # via pytest-cov
cryptography==43.0.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
defusedxml==0.7.1
    # via
    #   httpie
    #   semantic-kernel
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via
    #   -r .\requirements.in
    #   pymongo
docopt==0.6.2
    # via -r .\requirements.in
faker==26.1.0
    # via -r .\requirements.in
fastapi==0.112.0
    # via
    #   -r .\requirements.in
    #   fastapi-msal
fastapi-msal==2.1.4
    # via -r .\requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpie==3.2.3
    # via -r .\requirements.in
httpx==0.27.0
    # via
    #   -r .\requirements.in
    #   openai
hypercorn==0.17.3
    # via -r .\requirements.in
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via
    #   azure-storage-blob
    #   openapi-core
    #   rdflib
itsdangerous==2.2.0
    # via -r .\requirements.in
jinja2==3.1.4
    # via -r .\requirements.in
jsonschema==4.23.0
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.3
    # via openapi-spec-validator
jsonschema-spec==0.2.4
    # via openapi-core
jsonschema-specifications==2023.7.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via openapi-core
motor==3.5.1
    # via semantic-kernel
msal==1.30.0
    # via
    #   azure-identity
    #   fastapi-msal
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
multidict==6.0.5
    # via
    #   -r .\requirements.in
    #   aiohttp
    #   httpie
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.26.4
    # via
    #   pandas
    #   semantic-kernel
openai==1.38.0
    # via
    #   -r .\requirements.in
    #   semantic-kernel
openapi-core==0.18.2
    # via semantic-kernel
openapi-schema-validator==0.6.2
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi-core
packaging==24.1
    # via
    #   black
    #   prance
    #   pytest
pandas==2.2.2
    # via -r .\requirements.in
parse==1.20.2
    # via openapi-core
pathable==0.4.3
    # via
    #   jsonschema-path
    #   jsonschema-spec
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
portalocker==2.10.1
    # via msal-extensions
prance==23.6.21.0
    # via semantic-kernel
priority==2.0.0
    # via hypercorn
psutil==6.0.0
    # via -r .\requirements.in
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   fastapi
    #   fastapi-msal
    #   openai
    #   pydantic-settings
    #   semantic-kernel
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.4.0
    # via fastapi-msal
pygments==2.18.0
    # via
    #   httpie
    #   rich
pyjwt[crypto]==2.9.0
    # via
    #   msal
    #   pyjwt
pymongo==4.8.0
    # via
    #   -r .\requirements.in
    #   motor
pyparsing==3.1.2
    # via rdflib
pysocks==1.7.1
    # via requests
pytest==7.3.2
    # via
    #   -r .\requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.23.8
    # via -r .\requirements.in
pytest-cov==5.0.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   pandas
python-dotenv==1.0.1
    # via
    #   -r .\requirements.in
    #   pydantic-settings
    #   semantic-kernel
python-multipart==0.0.9
    # via -r .\requirements.in
pytz==2024.1
    # via pandas
pywin32==306
    # via portalocker
pyyaml==6.0.1
    # via
    #   jsonschema-path
    #   jsonschema-spec
rdflib==7.0.0
    # via -r .\requirements.in
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-spec
    #   jsonschema-specifications
regex==2023.12.25
    # via
    #   semantic-kernel
    #   tiktoken
requests[socks]==2.31.0
    # via
    #   -r .\requirements.in
    #   azure-core
    #   httpie
    #   jsonschema-path
    #   jsonschema-spec
    #   msal
    #   prance
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via httpie
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==13.7.1
    # via httpie
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via prance
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
semantic-kernel==0.9.1b1
    # via -r .\requirements.in
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   prance
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
starlette==0.37.2
    # via
    #   fastapi
    #   fastapi-msal
tiktoken==0.7.0
    # via -r .\requirements.in
tqdm==4.66.4
    # via openai
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   azure-storage-blob
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
uvicorn==0.30.5
    # via -r .\requirements.in
werkzeug==3.0.3
    # via openapi-core
wsproto==1.2.0
    # via hypercorn
xmlformatter==0.2.6
    # via -r .\requirements.in
yarl==1.9.4
    # via aiohttp
